% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GwasLocusScores.R
\name{GwasLocusScores}
\alias{GwasLocusScores}
\title{GwasLocusScores}
\value{
nothing

a data.frame

a data.frame

a data.frame

a data.frame
}
\description{
assess RNA-binding protein binding sites, their structure and motif matches
}
\examples{
  gls <- GwasLocusScores("rs4575096")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rbp}}{the gene symbol name of the protein}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rbp}}{the gene symbol name of the protein}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GwasLocusScores$new()}}
\item \href{#method-set.eqtls}{\code{GwasLocusScores$set.eqtls()}}
\item \href{#method-get.tbl.eqtl}{\code{GwasLocusScores$get.tbl.eqtl()}}
\item \href{#method-breakMotifsAtEQTLs}{\code{GwasLocusScores$breakMotifsAtEQTLs()}}
\item \href{#method-createTrenaMultiScoreTable}{\code{GwasLocusScores$createTrenaMultiScoreTable()}}
\item \href{#method-runTrena}{\code{GwasLocusScores$runTrena()}}
\item \href{#method-getADVX}{\code{GwasLocusScores$getADVX()}}
\item \href{#method-getETX}{\code{GwasLocusScores$getETX()}}
\item \href{#method-scoreEQTLs}{\code{GwasLocusScores$scoreEQTLs()}}
\item \href{#method-clone}{\code{GwasLocusScores$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$new(tag.snp, chrom, start.loc, end.loc, targetGene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag.snp}}{character, an rsid identifying the locus}

\item{\code{chrom}}{character, e.g., "chr1"}

\item{\code{start.loc}}{numeric the start of the genomic region of interest}

\item{\code{end.loc}}{numeric the end of the genomic region of interest}

\item{\code{targetGene}}{character, e.g. "NDUFS2"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of the GwasLocusScores class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.eqtls"></a>}}
\if{latex}{\out{\hypertarget{method-set.eqtls}{}}}
\subsection{Method \code{set.eqtls()}}{
specifies the eqtls to be used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$set.eqtls(tbl.eqtls)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.tbl.eqtl"></a>}}
\if{latex}{\out{\hypertarget{method-get.tbl.eqtl}{}}}
\subsection{Method \code{get.tbl.eqtl()}}{
extract all previously obtained eQTLs at or above the specified pvalue threshold
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$get.tbl.eqtl(pvalue.cutoff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pvalue.cutoff}}{numeric, e.g., 1e-4}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breakMotifsAtEQTLs"></a>}}
\if{latex}{\out{\hypertarget{method-breakMotifsAtEQTLs}{}}}
\subsection{Method \code{breakMotifsAtEQTLs()}}{
discover which motifs are broken by the eQTL variants, for specified gene and pval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$breakMotifsAtEQTLs(
  targetGene,
  pvalue.cutoff = NA,
  TFs.preselected = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetGene}}{character, a gene symbol in the eQTL table}

\item{\code{pvalue.cutoff}}{numeric, e.g., 1e-4}

\item{\code{TFs.preselected}}{character, e.g., c("SOX21", "ZEB1")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createTrenaMultiScoreTable"></a>}}
\if{latex}{\out{\hypertarget{method-createTrenaMultiScoreTable}{}}}
\subsection{Method \code{createTrenaMultiScoreTable()}}{
use FIMO and open chromatin, and TrenaMultiScore capabilities to create a tbl.tms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$createTrenaMultiScoreTable(
  trenaProject,
  tbl.fimo,
  tbl.oc,
  mtx.rna = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trenaProject}}{a concrete TrenaProject object, e.g., TrenaProjectAD()}

\item{\code{tbl.fimo}}{data.frame}

\item{\code{tbl.oc}}{data.frame}

\item{\code{mtx.rna}}{matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runTrena"></a>}}
\if{latex}{\out{\hypertarget{method-runTrena}{}}}
\subsection{Method \code{runTrena()}}{
given tfs, implicit target gene and an expression matrix, run trena
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$runTrena(tfs, mtx.rna, tbl.tms.final = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tfs}}{character, a vector of transcription factor geneSymbol names}

\item{\code{mtx.rna}}{matrix}

\item{\code{tbl.tms.final}}{data.frame, a TrenaMultiScore table from which the tfs came}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getADVX"></a>}}
\if{latex}{\out{\hypertarget{method-getADVX}{}}}
\subsection{Method \code{getADVX()}}{
access to the ADvariantExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$getADVX()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getETX"></a>}}
\if{latex}{\out{\hypertarget{method-getETX}{}}}
\subsection{Method \code{getETX()}}{
access to the EndophenotypeExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$getETX()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoreEQTLs"></a>}}
\if{latex}{\out{\hypertarget{method-scoreEQTLs}{}}}
\subsection{Method \code{scoreEQTLs()}}{
calculates the aggregate impact of broken motifs for TFs in the trena model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$scoreEQTLs(tbl.trena, tbl.breaks, tbl.eqtl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.trena}}{data.frame}

\item{\code{tbl.breaks}}{data.frame from motifbreakR}

\item{\code{tbl.eqtl}}{data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
